refactor(mcp): extract wiki_status collector into status.ts
